Rename relaying to relay
